ci: Add workflow to automate API reference sync on new version tags
The API reference generated in this repository is published via the
"docs" repository for the project. This means that we regularly need to
synchronize the document.

Instead of manually copying the API reference and creating a new Pull
Request in the "docs" repo, let's add a workflow to automate this task.
The workflow runs each time a new version tag is created in the "fabric"
repository.

Signed-off-by: Quentin Monnet <qmo@qmon.net>