ci: Update push-api-docs.yaml workflow to use a GitHub App token
Let's use the create-github-app-token Action to generate a GitHub App
token for the 'docs' repository, so that we can create a Pull Request
there, to sync up the API reference.

Comparing to a personal access token, this solution has the advantage of
making the GitHub App, rather than the contributor owning the token, the
author of the generated Pull Request.

Link: https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#authenticating-with-github-app-generated-tokens
Link: https://github.com/actions/create-github-app-token#usage
Signed-off-by: Quentin Monnet <qmo@qmon.net>