fix(agent): global staticExternal route adv
for static externals not bound to a VPC, we were adding their
subnets to a prefix list in the attached leaf and allowing them
to be advertised. However those routes will die at the spines
unless we also allow advertisment there. To do that, we must
pass the static external connection info to spines too (if they
are global) and populate prefix lists there as well.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>