fix(agent): externals as VPC
add also the IRB VLAN, else routes from the external are not going
to be advertised over BGP EVPN.

for the external outbound route-map, pre-emptively allow any prefix
belonging to the external's ipv4namespace, as opposed to selectively
adding VPC prefixes to a prefix-list depending on the external
peering. this way, routes learned via the gateway will be advertised
to the external, which allows traffic to be routed back to the VPC.

move externals VNIs and IRB VLANs before the VPCs, as they share
the same space and are less likely to change dynamically. This way
we will not see a different external VNI/VLAN every time we create
or delete a VPC.

make the catalog errors warnings for now, to allow agent upgrade.
these should be made into full blown errors eventually.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>