fix(agent): externals as VPC
add also the IRB VLAN, else routes from the external are not going
to be advertised over BGP EVPN.

move externals VNIs and IRB VLANs before the VPCs, as they share
the same space and are less likely to change dynamically. This way
we will not see a different external VNI/VLAN every time we create
or delete a VPC.

make the catalog errors warnings for now, to allow agent upgrade.
these should be made into full blown errors eventually.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>