feat: L2 external API
introduces an L2 external as a subtype of the existing BGP
speaker one. Ideally we would move the existing BGP params
in a sub-struct, but for now we want to avoid having to deal
with non backward-compatible changes.

this will need an equivalent L2 externalAttachment in order
to work; I will introduce it in a separate commit for ease
of review.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>