feat: L2 external API
introduces an L2 external as a subtype of the existing BGP
speaker one. Ideally we would move the existing BGP params
in a sub-struct, but for now we want to avoid having to deal
with non backward-compatible changes. similarly, adds a new
L2 external attachment subtype.

note that we do not need the static ARP config handler with
this solution, but I've kept it anyway as we might want to
use it in the future for some other feature.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>