chore(agent): adv. all VTEPs on leaves
regardless of fabric or mesh links. this allows us to
e.g. connect gateways to leaves in spine-leaf topologies,
whereas right now the VTEP of the gateway would not be
advertised by the leaf attached to it. The only
downside of this is that we are advertising some routes
which will be rejected by our peers, but the configuration
is simpler, and the alternative (adding the gateway VTEP
to the prefix list of advertised routes) is quite convoluted,
since that VTEP is only present on the gateway object, not
on the connection.

Signed-off-by: Emanuele Di Pascale <emanuele@githedgehog.com>